\name{awaptools-package}
\alias{awaptools-package}
\alias{awaptools}
\docType{package}
\title{
Functions to assist download and format of Australian Water Availability Grids
}
\description{
Various functions to assist download and format of Australian Water Availability Grids
}
\details{
\tabular{ll}{
Package: \tab awaptools\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2013-05-14\cr
License: \tab GPL (>= 2)\cr
}

}
\author{
ivanhanigan and ian szarka

Maintainer: Who to complain to ivan.hanigan@anu.edu.au

}
\references{

}

\keyword{ package }
\seealso{

}
\examples{

}
